+++ src/0_texture_init/texture.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/27 20:18:50 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/27 21:21:36 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/27 22:03:55 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,16 +35,28 @@ void	load_static_textures(void)
 	game()->textures.air_blight_death.frame1 = load_texture("assets/entities/blights/air/death1.png");
 	game()->textures.air_blight_death.frame2 = load_texture("assets/entities/blights/air/death2.png");
 	game()->textures.air_blight_death.frame3 = load_texture("assets/entities/blights/air/death3.png");
+	game()->textures.air_blight_attack.frame1 = load_texture("assets/entities/blights/air/attack1.png");
+	game()->textures.air_blight_attack.frame2 = load_texture("assets/entities/blights/air/attack2.png");
+	game()->textures.air_blight_attack.frame3 = load_texture("assets/entities/blights/air/attack3.png");
 	game()->textures.fire_blight_idle = load_texture("assets/entities/blights/fire/idle.png");
 	game()->textures.fire_blight_death.frame1 = load_texture("assets/entities/blights/fire/death1.png");
 	game()->textures.fire_blight_death.frame2 = load_texture("assets/entities/blights/fire/death2.png");
 	game()->textures.fire_blight_death.frame3 = load_texture("assets/entities/blights/fire/death3.png");
+	game()->textures.fire_blight_attack.frame1 = load_texture("assets/entities/blights/fire/attack1.png");
+	game()->textures.fire_blight_attack.frame2 = load_texture("assets/entities/blights/fire/attack2.png");
+	game()->textures.fire_blight_attack.frame3 = load_texture("assets/entities/blights/fire/attack3.png");
 	game()->textures.water_blight_idle = load_texture("assets/entities/blights/water/idle.png");
 	game()->textures.water_blight_death.frame1 = load_texture("assets/entities/blights/water/death1.png");
 	game()->textures.water_blight_death.frame2 = load_texture("assets/entities/blights/water/death2.png");
 	game()->textures.water_blight_death.frame3 = load_texture("assets/entities/blights/water/death3.png");
+	game()->textures.water_blight_attack.frame1 = load_texture("assets/entities/blights/water/attack1.png");
+	game()->textures.water_blight_attack.frame2 = load_texture("assets/entities/blights/water/attack2.png");
+	game()->textures.water_blight_attack.frame3 = load_texture("assets/entities/blights/water/attack3.png");
 	game()->textures.earth_blight_idle = load_texture("assets/entities/blights/earth/idle.png");
 	game()->textures.earth_blight_death.frame1 = load_texture("assets/entities/blights/earth/death1.png");
 	game()->textures.earth_blight_death.frame2 = load_texture("assets/entities/blights/earth/death2.png");
 	game()->textures.earth_blight_death.frame3 = load_texture("assets/entities/blights/earth/death3.png");
+	game()->textures.earth_blight_attack.frame1 = load_texture("assets/entities/blights/earth/attack1.png");
+	game()->textures.earth_blight_attack.frame2 = load_texture("assets/entities/blights/earth/attack2.png");
+	game()->textures.earth_blight_attack.frame3 = load_texture("assets/entities/blights/earth/attack3.png");
 }