+++ src/3_game_loop/a_game_logic/entities/entities.c
@@ -6,11 +6,25 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 11:47:39 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/27 15:47:20 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/27 22:07:35 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../include/cub3d.h"
+
+static inline bool	check_colliding(t_entity *ent1, t_entity *ent2)
+{
+	t_vec2	pos1;
+	t_vec2	pos2;
+
+	ent1->transform.rot = scale_vector(ent1->transform.rot, 1);
+	pos1.x = ent1->transform.pos.x + ent1->transform.rot.x * 0.5;
+	pos1.y = ent1->transform.pos.y + ent1->transform.rot.y * 0.5;
+	ent2->transform.rot = scale_vector(ent2->transform.rot, 1);
+	pos2.x = ent2->transform.pos.x + ent2->transform.rot.x * 0.5;
+	pos2.y = ent2->transform.pos.y + ent2->transform.rot.y * 0.5;
+	return (pos_distance(pos1, pos2) < DEFAULT_COLLISION_DISTANCE);
+}
 
 void	collide_entities(void)
 {
@@ -25,7 +39,7 @@ void	collide_entities(void)
 		{
 			if (ntt1 != ntt2)
 			{
-				if (pos_distance(((t_entity *)ntt1->content)->transform.pos, ((t_entity *)ntt2->content)->transform.pos) < DEFAULT_COLLISION_DISTANCE)
+				if (check_colliding((t_entity *)ntt1->content, (t_entity *)ntt2->content))
 				{
 					if (((t_entity *)ntt1->content)->on_collision)
 						((t_entity *)ntt1->content)->on_collision((t_entity *)ntt1->content, (t_entity *)ntt2->content);